% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{hb_delete_asset}
\alias{hb_delete_asset}
\title{Delete an asset}
\usage{
hb_delete_asset(client, url, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{url}{Asset URL (as returned by SeaTable in a file/image cell).}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Invisibly returns the client.
}
\description{
Delete an asset
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_delete_asset(client, "https://server/path/to/file.pdf")
\dontshow{\}) # examplesIf}
}
\seealso{
Other files:
\code{\link[=hb_attach_file]{hb_attach_file()}},
\code{\link[=hb_download_file]{hb_download_file()}},
\code{\link[=hb_upload_file]{hb_upload_file()}}
}
\concept{files}
