% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{hb_delete_select_option}
\alias{hb_delete_select_option}
\title{Delete a single-select option}
\usage{
hb_delete_select_option(client, table, name, option, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Table name.}

\item{name}{Column name.}

\item{option}{Option name to add.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Invisibly returns the client.
}
\description{
Delete a single-select option
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_delete_select_option(client, "Samples", "Status", "Old")
\dontshow{\}) # examplesIf}
}
\seealso{
Other columns:
\code{\link[=hb_add_column]{hb_add_column()}},
\code{\link[=hb_add_columns]{hb_add_columns()}},
\code{\link[=hb_add_select_option]{hb_add_select_option()}},
\code{\link[=hb_delete_column]{hb_delete_column()}},
\code{\link[=hb_list_columns]{hb_list_columns()}},
\code{\link[=hb_update_column]{hb_update_column()}},
\code{\link[=hb_update_select_option]{hb_update_select_option()}}
}
\concept{columns}
