% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{hb_delete_view}
\alias{hb_delete_view}
\title{Delete a view}
\usage{
hb_delete_view(client, table, view, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Table name.}

\item{view}{View name.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Invisibly returns the client.
}
\description{
Delete a view
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_delete_view(client, "Samples", "Old")
\dontshow{\}) # examplesIf}
}
\seealso{
Other views:
\code{\link[=hb_create_view]{hb_create_view()}},
\code{\link[=hb_get_view]{hb_get_view()}},
\code{\link[=hb_list_views.harbour_dtable]{hb_list_views.harbour_dtable()}},
\code{\link[=hb_update_view]{hb_update_view()}}
}
\concept{views}
