% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{hb_get_view}
\alias{hb_get_view}
\title{Get a view's settings}
\usage{
hb_get_view(client, table, view, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Table name.}

\item{view}{View name.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A 1-row tibble with the view's name, type and filters/sorts as
list-columns.
}
\description{
Get a view's settings
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_get_view(client, "Samples", "Default")
\dontshow{\}) # examplesIf}
}
\seealso{
Other views:
\code{\link[=hb_create_view]{hb_create_view()}},
\code{\link[=hb_delete_view]{hb_delete_view()}},
\code{\link[=hb_list_views]{hb_list_views()}},
\code{\link[=hb_update_view]{hb_update_view()}}
}
\concept{views}
