% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{hb_server_info}
\alias{hb_server_info}
\title{Server information}
\usage{
hb_server_info(client, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A one-row tibble with columns \code{server} (chr), \code{version} (chr) and
\code{edition} (chr) where reported.
}
\description{
Returns the SeaTable server's reported version and basic info. This
endpoint is unauthenticated.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_server_info(client)
\dontshow{\}) # examplesIf}
}
\seealso{
Other client:
\code{\link[=hb_check_credentials]{hb_check_credentials()}},
\code{\link[=hb_client]{hb_client()}},
\code{\link[=hb_ping]{hb_ping()}},
\code{\link[=is_harbour_client]{is_harbour_client()}}
}
\concept{client}
