% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{hb_update_view}
\alias{hb_update_view}
\title{Update a view}
\usage{
hb_update_view(client, table, view, settings, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Table name.}

\item{view}{New view name.}

\item{settings}{Optional list of view settings.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Invisibly returns the client.
}
\description{
Update a view
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_update_view(client, "Samples", "Active", list(filter_conjunction = "And"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other views:
\code{\link[=hb_create_view]{hb_create_view()}},
\code{\link[=hb_delete_view]{hb_delete_view()}},
\code{\link[=hb_get_view]{hb_get_view()}},
\code{\link[=hb_list_views.harbour_dtable]{hb_list_views.harbour_dtable()}}
}
\concept{views}
