% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{is_harbour_client}
\alias{is_harbour_client}
\title{Test whether an object is a harbour client}
\usage{
is_harbour_client(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is a harbour client
}
\seealso{
Other client:
\code{\link[=hb_check_credentials]{hb_check_credentials()}},
\code{\link[=hb_client]{hb_client()}},
\code{\link[=hb_ping]{hb_ping()}},
\code{\link[=hb_server_info]{hb_server_info()}}
}
\concept{client}
