% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{is_harbour_dtable}
\alias{is_harbour_dtable}
\title{Test whether an object is a harbour dtable}
\usage{
is_harbour_dtable(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is a harbour dtable
}
\examples{
is_harbour_dtable(hb_dtable(Samples = data.frame(x = 1)))
}
\seealso{
Other dtable:
\code{\link[=as_tibble.harbour_dtable]{as_tibble.harbour_dtable()}},
\code{\link[=hb_asset_path]{hb_asset_path()}},
\code{\link[=hb_dtable]{hb_dtable()}},
\code{\link[=hb_read_csv]{hb_read_csv()}},
\code{\link[=hb_read_dtable]{hb_read_dtable()}},
\code{\link[=hb_read_xlsx]{hb_read_xlsx()}},
\code{\link[=hb_validate_dtable]{hb_validate_dtable()}},
\code{\link[=hb_write_csv]{hb_write_csv()}},
\code{\link[=hb_write_dtable]{hb_write_dtable()}},
\code{\link[=hb_write_xlsx]{hb_write_xlsx()}},
\code{\link[=length.harbour_dtable]{length.harbour_dtable()}},
\code{\link[=names.harbour_dtable]{names.harbour_dtable()}},
\code{\link[=print.harbour_dtable]{print.harbour_dtable()}},
\code{\link[=summary.harbour_dtable]{summary.harbour_dtable()}}
}
\concept{dtable}
