% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{is_harbour_metadata}
\alias{is_harbour_metadata}
\title{Test whether an object is harbour metadata}
\usage{
is_harbour_metadata(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is harbour metadata
}
\seealso{
Other metadata:
\code{\link[=as_tibble.harbour_metadata]{as_tibble.harbour_metadata()}},
\code{\link[=hb_column_types]{hb_column_types()}},
\code{\link[=hb_list_collaborators]{hb_list_collaborators()}},
\code{\link[=hb_list_tables.harbour_dtable]{hb_list_tables.harbour_dtable()}},
\code{\link[=hb_metadata]{hb_metadata()}},
\code{\link[=print.harbour_metadata]{print.harbour_metadata()}},
\code{\link[=summary.harbour_metadata]{summary.harbour_metadata()}}
}
\concept{metadata}
