% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-methods.R
\name{length.harbour_dtable}
\alias{length.harbour_dtable}
\title{Number of tables in a dtable}
\usage{
\method{length}{harbour_dtable}(x)
}
\arguments{
\item{x}{A \code{harbour_dtable}.}
}
\value{
A single integer.
}
\description{
Number of tables in a dtable
}
\examples{
length(hb_dtable(Samples = data.frame(x = 1)))
}
\seealso{
Other dtable:
\code{\link[=as_tibble.harbour_dtable]{as_tibble.harbour_dtable()}},
\code{\link[=hb_asset_path]{hb_asset_path()}},
\code{\link[=hb_dtable]{hb_dtable()}},
\code{\link[=hb_read_csv]{hb_read_csv()}},
\code{\link[=hb_read_dtable]{hb_read_dtable()}},
\code{\link[=hb_read_xlsx]{hb_read_xlsx()}},
\code{\link[=hb_validate_dtable]{hb_validate_dtable()}},
\code{\link[=hb_write_csv]{hb_write_csv()}},
\code{\link[=hb_write_dtable]{hb_write_dtable()}},
\code{\link[=hb_write_xlsx]{hb_write_xlsx()}},
\code{\link[=is_harbour_dtable]{is_harbour_dtable()}},
\code{\link[=names.harbour_dtable]{names.harbour_dtable()}},
\code{\link[=print.harbour_dtable]{print.harbour_dtable()}},
\code{\link[=summary.harbour_dtable]{summary.harbour_dtable()}}
}
\concept{dtable}
