% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-methods.R
\name{print.harbour_dtable}
\alias{print.harbour_dtable}
\alias{format.harbour_dtable}
\title{Print a harbour dtable}
\usage{
\method{print}{harbour_dtable}(x, ...)

\method{format}{harbour_dtable}(x, ...)
}
\arguments{
\item{x}{A \code{harbour_dtable}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.

A character vector, one element per table.
}
\description{
Print a harbour dtable
}
\examples{
print(hb_dtable(Samples = data.frame(x = 1)))
}
\seealso{
Other dtable:
\code{\link[=as_tibble.harbour_dtable]{as_tibble.harbour_dtable()}},
\code{\link[=hb_asset_path]{hb_asset_path()}},
\code{\link[=hb_dtable]{hb_dtable()}},
\code{\link[=hb_read_csv]{hb_read_csv()}},
\code{\link[=hb_read_dtable]{hb_read_dtable()}},
\code{\link[=hb_read_xlsx]{hb_read_xlsx()}},
\code{\link[=hb_validate_dtable]{hb_validate_dtable()}},
\code{\link[=hb_write_csv]{hb_write_csv()}},
\code{\link[=hb_write_dtable]{hb_write_dtable()}},
\code{\link[=hb_write_xlsx]{hb_write_xlsx()}},
\code{\link[=is_harbour_dtable]{is_harbour_dtable()}},
\code{\link[=length.harbour_dtable]{length.harbour_dtable()}},
\code{\link[=names.harbour_dtable]{names.harbour_dtable()}},
\code{\link[=summary.harbour_dtable]{summary.harbour_dtable()}}
}
\concept{dtable}
