% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{as_tibble.harbour_metadata}
\alias{as_tibble.harbour_metadata}
\title{Coerce harbour metadata to a tibble}
\usage{
\method{as_tibble}{harbour_metadata}(x, ...)
}
\arguments{
\item{x}{A \code{harbour_metadata} object.}

\item{...}{Unused.}
}
\value{
A tibble with one row per table; see \code{\link[=hb_list_tables]{hb_list_tables()}}.
}
\description{
Coerce harbour metadata to a tibble
}
\seealso{
Other metadata:
\code{\link[=hb_column_types]{hb_column_types()}},
\code{\link[=hb_list_collaborators]{hb_list_collaborators()}},
\code{\link[=hb_list_tables]{hb_list_tables()}},
\code{\link[=hb_metadata]{hb_metadata()}},
\code{\link[=is_harbour_metadata]{is_harbour_metadata()}},
\code{\link[=print.harbour_metadata]{print.harbour_metadata()}},
\code{\link[=summary.harbour_metadata]{summary.harbour_metadata()}}
}
\concept{metadata}
