% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbouR-package.R
\docType{package}
\name{harbouR-package}
\alias{harbouR}
\alias{harbouR-package}
\title{harbouR: R Client for SeaTable Collaborative Databases}
\description{
An unofficial R client for the SeaTable REST API. Connect to a SeaTable
server, read and write rows as tidy tibbles, manage tables/columns/views,
upload and attach files, and explore bases interactively with the bundled
Shiny app.
}
\details{
harbouR was shaped by the needs, field testing and data of Paul
Elsinghorst, Wiebke Derz, Matthias Ring, Gerhard Achatz and Vinzent
Forstmeier.

SeaTable is a trademark of SeaTable GmbH; harbouR is not affiliated with
or endorsed by SeaTable GmbH.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CTTIR/harbouR}
  \item \url{https://cttir.github.io/harbouR/}
  \item Report bugs at \url{https://github.com/CTTIR/harbouR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Raban Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Paul Elsinghorst [contributor]
  \item Wiebke Derz [contributor]
  \item Matthias Ring [contributor]
  \item Gerhard Achatz [contributor]
  \item Vinzent Forstmeier [contributor]
}

}
\keyword{internal}
