% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbouR-package.R
\docType{package}
\name{harbouR-package}
\alias{harbouR}
\alias{harbouR-package}
\title{harbouR: R Client for SeaTable Collaborative Databases}
\description{
An unofficial R client for the SeaTable REST API. Connect to a SeaTable
server, read and write rows as tidy tibbles, manage tables/columns/views,
upload and attach files, and explore bases interactively with the bundled
Shiny app.
}
\details{
SeaTable is a trademark of SeaTable GmbH; harbouR is not affiliated with
or endorsed by SeaTable GmbH.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-heller/harbouR}
  \item \url{https://r-heller.github.io/harbouR/}
  \item Report bugs at \url{https://github.com/r-heller/harbouR/issues}
}

}
\author{
\strong{Maintainer}: R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

Authors:
\itemize{
  \item R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})
}

}
\keyword{internal}
