% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hb_delete_table}
\alias{hb_delete_table}
\title{Delete a table}
\usage{
hb_delete_table(client, table, call = rlang::caller_env())
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Name of the table to delete.}

\item{call}{Internal: error-propagation env.}
}
\value{
Invisibly returns the client.
}
\description{
Delete a table
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_delete_table(client, "DropMe")
\dontshow{\}) # examplesIf}
}
\seealso{
Other tables:
\code{\link[=hb_create_table]{hb_create_table()}},
\code{\link[=hb_duplicate_table]{hb_duplicate_table()}},
\code{\link[=hb_rename_table]{hb_rename_table()}}
}
\concept{tables}
