% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{hb_get_row}
\alias{hb_get_row}
\title{Get a single row by ID}
\usage{
hb_get_row(client, table, row_id, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Name of the table.}

\item{row_id}{The SeaTable row identifier.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A 1-row tibble, or a 0-row tibble if the row is not found.
}
\description{
Get a single row by ID
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_get_row(client, "Samples", "abc123")
\dontshow{\}) # examplesIf}
}
\seealso{
Other rows:
\code{\link[=hb_append_rows]{hb_append_rows()}},
\code{\link[=hb_delete_rows]{hb_delete_rows()}},
\code{\link[=hb_lock_rows]{hb_lock_rows()}},
\code{\link[=hb_query]{hb_query()}},
\code{\link[=hb_read_table]{hb_read_table()}},
\code{\link[=hb_unlock_rows]{hb_unlock_rows()}},
\code{\link[=hb_update_rows]{hb_update_rows()}}
}
\concept{rows}
