% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{hb_list_columns}
\alias{hb_list_columns}
\title{List columns of a table}
\usage{
hb_list_columns(client, table, call = rlang::caller_env())
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Table name.}

\item{call}{Internal: error-propagation env.}
}
\value{
A tibble with columns \code{name} (chr), \code{type} (chr), \code{key} (chr) and
\code{editable} (lgl).
}
\description{
List columns of a table
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_list_columns(client, "Samples")
\dontshow{\}) # examplesIf}
}
\seealso{
Other columns:
\code{\link[=hb_add_column]{hb_add_column()}},
\code{\link[=hb_add_columns]{hb_add_columns()}},
\code{\link[=hb_add_select_option]{hb_add_select_option()}},
\code{\link[=hb_delete_column]{hb_delete_column()}},
\code{\link[=hb_delete_select_option]{hb_delete_select_option()}},
\code{\link[=hb_update_column]{hb_update_column()}},
\code{\link[=hb_update_select_option]{hb_update_select_option()}}
}
\concept{columns}
