% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{hb_list_tables}
\alias{hb_list_tables}
\title{List the tables in a base}
\usage{
hb_list_tables(client, refresh = FALSE, call = rlang::caller_env())
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{refresh}{Logical; refetch metadata even if cached. Default \code{FALSE}.}

\item{call}{Internal: error-propagation env.}
}
\value{
A tibble with one row per table and columns
\code{name} (chr), \code{n_rows} (int), \code{n_columns} (int), \code{n_views} (int).
}
\description{
List the tables in a base
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_list_tables(client)
\dontshow{\}) # examplesIf}
}
\seealso{
Other metadata:
\code{\link[=as_tibble.harbour_metadata]{as_tibble.harbour_metadata()}},
\code{\link[=hb_column_types]{hb_column_types()}},
\code{\link[=hb_list_collaborators]{hb_list_collaborators()}},
\code{\link[=hb_metadata]{hb_metadata()}},
\code{\link[=is_harbour_metadata]{is_harbour_metadata()}},
\code{\link[=print.harbour_metadata]{print.harbour_metadata()}},
\code{\link[=summary.harbour_metadata]{summary.harbour_metadata()}}
}
\concept{metadata}
