% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{hb_list_views}
\alias{hb_list_views}
\title{List views of a table}
\usage{
hb_list_views(client, table, call = rlang::caller_env())
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Table name.}

\item{call}{Internal: error-propagation env.}
}
\value{
A tibble with columns \code{name} (chr), \code{type} (chr) and \code{is_default}
(lgl). Zero rows if no views exist.
}
\description{
List views of a table
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_list_views(client, "Samples")
\dontshow{\}) # examplesIf}
}
\seealso{
Other views:
\code{\link[=hb_create_view]{hb_create_view()}},
\code{\link[=hb_delete_view]{hb_delete_view()}},
\code{\link[=hb_get_view]{hb_get_view()}},
\code{\link[=hb_update_view]{hb_update_view()}}
}
\concept{views}
