% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{hb_query}
\alias{hb_query}
\title{Run a SeaTable SQL query}
\usage{
hb_query(client, sql, call = rlang::caller_env())
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{sql}{SeaTable SQL query string.}

\item{call}{Internal: error-propagation env.}
}
\value{
A tibble. Always a tibble, even when the query returns no rows.
}
\description{
Run a SeaTable SQL query
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_query(client, "select * from Samples limit 5")
\dontshow{\}) # examplesIf}
}
\seealso{
Other rows:
\code{\link[=hb_append_rows]{hb_append_rows()}},
\code{\link[=hb_delete_rows]{hb_delete_rows()}},
\code{\link[=hb_get_row]{hb_get_row()}},
\code{\link[=hb_lock_rows]{hb_lock_rows()}},
\code{\link[=hb_read_table]{hb_read_table()}},
\code{\link[=hb_unlock_rows]{hb_unlock_rows()}},
\code{\link[=hb_update_rows]{hb_update_rows()}}
}
\concept{rows}
