% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hb_rename_table}
\alias{hb_rename_table}
\title{Rename a table}
\usage{
hb_rename_table(client, table, new_name, call = rlang::caller_env())
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Current table name.}

\item{new_name}{New table name.}

\item{call}{Internal: error-propagation env.}
}
\value{
Invisibly returns the client.
}
\description{
Rename a table
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_rename_table(client, "Old", "New")
\dontshow{\}) # examplesIf}
}
\seealso{
Other tables:
\code{\link[=hb_create_table]{hb_create_table()}},
\code{\link[=hb_delete_table]{hb_delete_table()}},
\code{\link[=hb_duplicate_table]{hb_duplicate_table()}}
}
\concept{tables}
