% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{hb_unlock_rows}
\alias{hb_unlock_rows}
\title{Unlock rows}
\usage{
hb_unlock_rows(client, table, row_ids, ...)
}
\arguments{
\item{client}{A \code{harbour_client}.}

\item{table}{Table name.}

\item{row_ids}{A character vector of row IDs to delete.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Invisibly returns the client.
}
\description{
Unlock rows
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
client <- hb_client()
hb_unlock_rows(client, "Samples", "abc")
\dontshow{\}) # examplesIf}
}
\seealso{
Other rows:
\code{\link[=hb_append_rows]{hb_append_rows()}},
\code{\link[=hb_delete_rows]{hb_delete_rows()}},
\code{\link[=hb_get_row]{hb_get_row()}},
\code{\link[=hb_lock_rows]{hb_lock_rows()}},
\code{\link[=hb_query]{hb_query()}},
\code{\link[=hb_read_table.harbour_dtable]{hb_read_table.harbour_dtable()}},
\code{\link[=hb_update_rows]{hb_update_rows()}}
}
\concept{rows}
