% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{print.harbour_metadata}
\alias{print.harbour_metadata}
\title{Print method for harbour metadata}
\usage{
\method{print}{harbour_metadata}(x, ...)
}
\arguments{
\item{x}{A \code{harbour_metadata} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for harbour metadata
}
\seealso{
Other metadata:
\code{\link[=as_tibble.harbour_metadata]{as_tibble.harbour_metadata()}},
\code{\link[=hb_column_types]{hb_column_types()}},
\code{\link[=hb_list_collaborators]{hb_list_collaborators()}},
\code{\link[=hb_list_tables]{hb_list_tables()}},
\code{\link[=hb_metadata]{hb_metadata()}},
\code{\link[=is_harbour_metadata]{is_harbour_metadata()}},
\code{\link[=summary.harbour_metadata]{summary.harbour_metadata()}}
}
\concept{metadata}
